Removes redundant http client
The http client was being stored in the Client struct, and initialized when
the client was constructed. This commit removes the http client from the
client struct and initializes it within the methods that require it. This
simplifies the client struct and removes the need to clone the http client
when cloning the client.